[XENOPS-794] Put AMP resource files on the test classpath without adding them to test resources.
When adding the amp resources to the test classpath by adding them to test resources, intellij recognizes them as test resources, which results in a slightly confusing project layout.
Instead, we can add the amp itself to the test classpath. This will still make all resources in the AMP available on the test classpath as resources at runtime.